% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow-types.R
\name{arrow-types}
\alias{arrow-types}
\alias{arrowIntType}
\alias{arrowType}
\alias{arrowTypeToName}
\alias{arrowTypeFromName}
\title{Arrow type helpers for the DuckDB package suite}
\usage{
arrowIntType(range)

arrowType(x)

arrowTypeToName(type)

arrowTypeFromName(name)
}
\arguments{
\item{range}{Length-two numeric vector \code{c(min, max)} for
\code{\link{arrowIntType}}.}

\item{x}{An R vector for \code{\link{arrowType}}.}

\item{type}{An \code{\link[arrow]{DataType}} or Arrow type name (character)
for \code{\link{arrowTypeToName}}.}

\item{name}{Length-one character Arrow type name (e.g. \code{"uint16"})
for \code{\link{arrowTypeFromName}}.}
}
\value{
\describe{
  \item{\code{arrowIntType}, \code{arrowType}}{
    An \code{\link[arrow]{DataType}}.}
  \item{\code{arrowTypeToName}}{
    A length-one character Arrow type name.}
  \item{\code{arrowTypeFromName}}{
    An \code{\link[arrow]{DataType}}.}
}
}
\description{
Utilities for narrowing integer types and converting between
\code{\link[arrow]{DataType}} objects and Arrow type names. Used by
\pkg{BiocDuckDB} and \pkg{DuckDBArray} for Parquet schema selection
and parallel-safe type passing.
}
\details{
\code{arrowIntType} selects the narrowest unsigned or signed Arrow integer
type that can represent \code{range}. \code{arrowType} infers an Arrow type
from an R vector, narrowing integers when possible.

\code{arrowTypeFromName} and \code{arrowTypeToName} implement a small
registry of scalar Arrow types used when passing type information to
BiocParallel workers (where \code{DataType} external pointers do not
serialize reliably).
}
\examples{
arrowIntType(c(0L, 10L))
arrowType(1:10L)
arrowTypeToName(arrow::uint16())
arrowTypeFromName("uint16")

}
\seealso{
\code{\link{parquet-io}},
\code{\link[BiocDuckDB]{writeParquet}},
\code{\link[DuckDBArray]{writeCoordArray}}
}
\author{
Patrick Aboyoun
}
