% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow-types.R
\name{arrowIntType}
\alias{arrowIntType}
\title{Narrowest Arrow integer type for a numeric range}
\usage{
arrowIntType(range)
}
\arguments{
\item{range}{Length-two numeric vector \code{c(min, max)}.}
}
\value{
An \code{\link[arrow]{DataType}}.
}
\description{
Selects the narrowest unsigned or signed Arrow integer type that can
represent \code{range}.
}
\author{
Patrick Aboyoun
}
