% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow-types.R
\name{arrowType}
\alias{arrowType}
\title{Infer Arrow type from R values}
\usage{
arrowType(x)
}
\arguments{
\item{x}{An R vector.}
}
\value{
An \code{\link[arrow]{DataType}}.
}
\description{
Infers an Arrow type from an R vector, narrowing integers when possible.
}
\author{
Patrick Aboyoun
}
