% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuckDBAtomicList-class.R
\name{DuckDBAtomicList-class}
\alias{DuckDBAtomicList-class}
\alias{DuckDBLogicalList-class}
\alias{DuckDBIntegerList-class}
\alias{DuckDBInteger64List-class}
\alias{DuckDBNumericList-class}
\alias{DuckDBCharacterList-class}
\alias{DuckDBFactorList-class}
\alias{elementNROWS,DuckDBAtomicList-method}
\alias{extractROWS,DuckDBAtomicList,ANY-method}
\alias{getListElement,DuckDBAtomicList-method}
\alias{head,DuckDBAtomicList-method}
\alias{tail,DuckDBAtomicList-method}
\alias{as.list,DuckDBAtomicList-method}
\alias{realize,DuckDBAtomicList-method}
\alias{show,DuckDBAtomicList-method}
\alias{showAsCell,DuckDBAtomicList-method}
\title{DuckDBAtomicList objects}
\description{
The DuckDBAtomicList class extends \linkS4class{AtomicList} and
\linkS4class{DuckDBColumn} to represent list columns extracted from a
\linkS4class{DuckDBDataFrame} object.
}
\section{Accessors}{

In the code snippets below, \code{x} is a DuckDBAtomicList object:
\describe{
  \item{\code{length(x)}:}{
    Get the number of list elements in \code{x}.
  }
  \item{\code{names(x)}:}{
    Get the names of the list elements of \code{x}.
  }
  \item{\code{elementNROWS(x)}:}{
    Get the length of each list element in \code{x}.
  }
  \item{\code{elementType(x)}:}{
    Get the type of the list elements; one of \code{"logical"},
    \code{"integer"}, \code{"integer64"}, \code{"numeric"},
    \code{"character"}, or \code{"factor"}.
  }
  \item{\code{dimtbls(x, drop = TRUE)}, \code{dimtbls(x) <- value}:}{
    Get or set the list of dimension tables used to define partitions for
    efficient queries. If \code{drop = TRUE}, then it returns a named
    \code{DataFrameList} object, else it returns an environment containing
    a \code{dimtbls} named \code{DataFrameList} element.
  }
}
}

\section{Coercion}{

\describe{
  \item{\code{as.list(x)}:}{
    Coerces \code{x} to a list.
  }
  \item{\code{realize(x, BACKEND = getAutoRealizationBackend())}:}{
    Realize an object into memory or on disk using the equivalent of
    \code{realize(as.list(x), BACKEND)}.
  }
}
}

\section{Subsetting}{

In the code snippets below, \code{x} is a DuckDBAtomicList object:
\describe{
  \item{\code{x[i]}:}{
    Returns a DuckDBAtomicList object containing the selected list elements.
  }
  \item{\code{x[[i]]}:}{
    Extracts the list element at position \code{i} as an atomic vector.
  }
  \item{\code{head(x, n = 6L)}:}{
    If \code{n} is non-negative, returns the first n list elements of \code{x}.
    If \code{n} is negative, returns all but the last \code{abs(n)} list
    elements of \code{x}.
  }
  \item{\code{tail(x, n = 6L)}:}{
    If \code{n} is non-negative, returns the last n list elements of \code{x}.
    If \code{n} is negative, returns all but the first \code{abs(n)} list
    elements of \code{x}.
  }
}
}

\seealso{
\itemize{
  \item \code{\link{DuckDBColumn-class}} for atomic columns
  \item \code{\link[IRanges]{AtomicList}} for the base class
  \item \code{\link[S4Vectors]{List}} for the List class hierarchy
}
}
\author{
Patrick Aboyoun
}
\keyword{classes}
\keyword{methods}
