% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuckDBDataFrame-package.R
\docType{package}
\name{DuckDBDataFrame-package}
\alias{DuckDBDataFrame-package}
\title{DuckDBDataFrame: DuckDB-Backed DataFrame and Table Structures}
\description{
Provides DuckDB-backed implementations of core tabular data structures including DuckDBTable, DuckDBDataFrame, and DuckDBColumn. These classes extend Bioconductor's S4Vectors framework to enable efficient, out-of-memory operations on large tabular datasets using DuckDB's analytical database engine. The package includes SQL function discovery, connection management, and support for list columns and embeddings stored in ARRAY[] columns.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Genentech/DuckDBDataFrame}
  \item Report bugs at \url{https://github.com/Genentech/DuckDBDataFrame/issues}
}

}
\author{
\strong{Maintainer}: Patrick Aboyoun \email{aboyounp@gene.com}

Authors:
\itemize{
  \item Patrick Aboyoun \email{aboyounp@gene.com}
}

Other contributors:
\itemize{
  \item Genentech, Inc. [copyright holder]
}

}
\keyword{internal}
